refactor: Rename file {keyboardShortcuts.ts -> keyEventUtils.ts}
